// types/form-data/brandingFormData.ts
export interface BrandingFormData {
	backgroundColor: string;
	discord: string;
	failureCallToActionButtonText: string;
	failureCallToActionText: string;
	failureCallToActionURL: string;
	fontColor: string;
	logoFileContent: string;
	logoFileMime: string;
	logoFileName: string;
	primaryColor: string;
	secondaryColor: string;
	secondaryFailureCallToActionButtonText: string;
	secondaryFailureCallToActionURL: string;
	secondarySuccessCallToActionButtonText: string;
	secondarySuccessCallToActionURL: string;
	successCallToActionButtonText: string;
	successCallToActionText: string;
	successCallToActionURL: string;
	telegram: string;
	websiteURL: string;
	x: string;
	youtube: string;
}